Add flow of adding and removing stuff from datablock.

+++ include/rowid.h
@@ -6,13 +6,13 @@
 class RowID
 {
 public:
-	RowID(int32_t blockNumber, int32_t position);
+	RowID(int16_t blockNumber, int16_t position);
 	~RowID();
-	int32_t getBlockNumber();
-	int32_t getPosition();
+	int16_t getBlockNumber();
+	int16_t getPosition();
 private:
-	int32_t blockNumber;
-	int32_t position;
+	int16_t blockNumber;
+	int16_t position;
 };
 
 #endif //ROWID_H
