+++ src/rowid.cpp
@@ -1,6 +1,6 @@
 #include "rowid.h"
 
-RowID::RowID(int32_t blockNumber, int32_t position)
+RowID::RowID(int16_t blockNumber, int16_t position)
 {
 	this->blockNumber = blockNumber;
 	this->position = position;
@@ -11,12 +11,12 @@ RowID::~RowID()
 
 }
 
-int32_t RowID::getBlockNumber()
+int16_t RowID::getBlockNumber()
 {
 	return blockNumber;
 }
 
-int32_t RowID::getPosition()
+int16_t RowID::getPosition()
 {
 	return position;
 }
